+++ src/components/NodeNeighbourhood.js
@@ -8,21 +8,16 @@ import { useTranslation } from 'react-i18next'
 import { InputNumber } from 'primereact/inputnumber'
 import { Button } from 'primereact/button'
 import actions from '../store/actions'
-import { SIDEBAR_VIEW_GRAPHS, SIDEBAR_VIEW_NEIGHBOURHOOD } from '../constants/views'
-import { ALGO_TYPE_NEIGHBOURHOOD } from '../constants/algorithms'
+import { SIDEBAR_VIEW_NEIGHBOURHOOD } from '../constants/views'
+import setNeighbourNodes from '../utils/setNeighbourNodes'
 
 const NodeNeighbourhood = ({
   setStoreState,
-  lastGraphIndex,
   classesFromApi,
   selectedNeighbourNode,
   updateGraphData,
-  triplesPerNode
 }) => {
   const { t } = useTranslation()
-
-  const newGraphIndex = lastGraphIndex + 1
-  const newCurrentGraph = `graph-${lastGraphIndex + 1}`
 
   const [separationDegree, setSeparationDegree] = useState(1)
 
@@ -48,11 +43,11 @@ const NodeNeighbourhood = ({
               <table>
                 <tbody>
                   <tr>
-                    <td>{t('label')}</td>
+                    <td className="bold">{t('label')}</td>
                     <td>{classesFromApi[selectedNeighbourNode].label}</td>
                   </tr>
                   <tr>
-                    <td>{t('id')}</td>
+                    <td className="bold">{t('id')}</td>
                     <td>{classesFromApi[selectedNeighbourNode].id}</td>
                   </tr>
                 </tbody>
@@ -81,25 +76,11 @@ const NodeNeighbourhood = ({
           icon="pi pi-chevron-right"
           iconPos="right"
           label={t('show')}
-          onClick={() => {
-            const selectedNodeId = classesFromApi[selectedNeighbourNode] ? classesFromApi[selectedNeighbourNode].id : ''
-            const label = `neighbourhood-${newCurrentGraph}`
-
-            const graphValue = {
-              label,
-              type: ALGO_TYPE_NEIGHBOURHOOD,
-              options: {
-                selectedNodeId,
-                separationDegree,
-                triplesPerNode
-              }
-            }
-
-            updateGraphData(newCurrentGraph, graphValue)
-            setStoreState('currentGraph', newCurrentGraph)
-            setStoreState('lastGraphIndex', newGraphIndex)
-            setStoreState('sidebarView', SIDEBAR_VIEW_GRAPHS)
-          }}
+          onClick={() => setNeighbourNodes({
+            separationDegree,
+            setStoreState,
+            updateGraphData
+          })}
         />
       </div>
     </>
@@ -108,27 +89,21 @@ const NodeNeighbourhood = ({
 
 NodeNeighbourhood.propTypes = {
   setStoreState: PropTypes.func.isRequired,
-  lastGraphIndex: PropTypes.number.isRequired,
   selectedNeighbourNode: PropTypes.string.isRequired,
   classesFromApi: PropTypes.shape().isRequired,
   updateGraphData: PropTypes.func.isRequired,
-  triplesPerNode: PropTypes.shape().isRequired,
 }
 
 const mapToProps = ({
   graphData,
   currentGraph,
-  lastGraphIndex,
   classesFromApi,
   selectedNeighbourNode,
-  triplesPerNode
 }) => ({
   graphData,
   currentGraph,
-  lastGraphIndex,
   classesFromApi,
   selectedNeighbourNode,
-  triplesPerNode
 })
 
 export default connect(