+++ src/store/index.js
@@ -35,6 +35,9 @@ const initialState = {
   exportFileName: 'network-graph',
   freeTextSelection: {},
   freeTextSelectedElement: '',
+  isShortestPathNodeSelectable: false,
+  shortestPathSelectedNodes: [],
+  shortestPathResults: [],
 
   // states to update at every view refresh
   nodesIdsToDisplay: [],
@@ -51,6 +54,8 @@ const initialState = {
   selectedNeighbourNode: '',
   isNeighbourNodeSelectable: false,
   highlightedNodes: [],
+  isNodeOverlay: false,
+  paths: [],
 
   // view data storage
   lastGraphIndex: 0,