+++ src/utils/serialiseNodesEdges/addNode.js
@@ -7,17 +7,21 @@ import { SELECTED_NODE_COLOR } from '../../constants/graph'
  * @param  {Array}    params.availableNodesList       Array of available nodes
  * @param  {Object}   params.availableNodesNormalised Normalised list of available nodes
  * @param  {Array}    params.highlightedNodes         Array of nodes IDs to highlight
+ * @param  {Boolean}  params.isNodeOverlay            Flag to make non-highlighted nodes transparent
  * @param  {String}   params.nodeId                   Node ID
  * @param  {Object}   params.nodeIdObject             Node ID object
+ * @param  {Array}    params.shortestPathNodes        Array of nodes IDs in shortest path
  * @return
  */
 const addNode = ({
   addedNodes,
   availableNodesList,
   availableNodesNormalised,
+  isNodeOverlay,
   highlightedNodes,
   nodeId,
   nodeIdObject,
+  shortestPathNodes
 }) => {
   if (!addedNodes.includes(nodeId)
   && nodeIdObject.label
@@ -29,6 +33,8 @@ const addNode = ({
       extendedNodeObject.color = {
         background: SELECTED_NODE_COLOR
       }
+    } else if (isNodeOverlay && !shortestPathNodes.includes(nodeId)) {
+      extendedNodeObject.opacity = 0.1
     }
 
     availableNodesNormalised[nodeId] = extendedNodeObject  // eslint-disable-line